pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fishing-assistant"
version = "1.0.0"
description = "Fishing Conditions Research Crew using CrewAI"
authors = [
    {name = "Rich Chow", email = "askqai@mitns.com"}
]
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "crewai>=0.157.0",
    "crewai-tools>=0.6.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "pydantic>=2.9.2",
    "pyyaml>=6.0.1",
    "ollama>=0.3.0",
    "pandas>=2.2.3",
    "litellm>=1.70.0",
]


[project.scripts]
fishing-assistant = "fishing_assistant.main:main"

[project.urls]
Homepage = "https://github.com/rchow93/fishing-assistant"
Repository = "https://github.com/rchow93/fishing-assistant"
Issues = "https://github.com/rchow93/fishing-assistant"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["fishing_assistant*"]

[tool.setuptools.package-data]
fishing_assistant = ["config/*.yaml"]
